pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "autotrader-bot"
version = "0.1.0"
description = "Continuous Learning Cryptocurrency Trading Bot"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Autotrader Development Team", email = "autotrader@example.com"}
]
keywords = ["cryptocurrency", "trading", "bot", "machine-learning", "tensorflow", "btc", "bitcoin"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Financial and Insurance Industry", 
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9", 
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Office/Business :: Financial :: Investment",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
requires-python = ">=3.8"
dependencies = [
    "tensorflow>=2.13.0,<3.0.0",
    "numpy>=1.21.0,<2.0.0",
    "pandas>=1.5.0,<3.0.0",
    "scikit-learn>=1.3.0,<2.0.0",
    "requests>=2.28.0,<3.0.0",
    "aiohttp>=3.8.0,<4.0.0",
    "websockets>=10.0,<12.0",
    "python-dotenv>=0.19.0,<2.0.0",
    "pyyaml>=6.0,<7.0",
    "pydantic>=1.10.0,<3.0.0",
    "structlog>=22.0.0,<24.0.0",
    "tenacity>=8.0.0,<9.0.0",
    "click>=8.0.0,<9.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0", 
    "pytest-asyncio>=0.21.0",
    "black>=22.0.0",
    "flake8>=5.0.0",
    "mypy>=1.0.0",
    "pre-commit>=2.20.0",
    "jupyter>=1.0.0",
    "notebook>=6.4.0",
    "ipywidgets>=8.0.0",
]
talib = [
    "talib-binary>=0.4.19",
]

[project.scripts]
autotrader = "autotrader.cli:main"
autotrader-bot = "autotrader.main:main"

[project.urls]
Homepage = "https://github.com/yourusername/autotrader-bot"
Repository = "https://github.com/yourusername/autotrader-bot"
Documentation = "https://autotrader-bot.readthedocs.io/"
"Bug Reports" = "https://github.com/yourusername/autotrader-bot/issues"

[tool.setuptools]
packages = ["autotrader"]

[tool.setuptools.package-data]
autotrader = ["config/*.yaml", "config/*.yml", "config/*.ini"]

# Black formatting
[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

# Pytest configuration
[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]
asyncio_mode = "auto"

# Coverage configuration
[tool.coverage.run]
source = ["autotrader"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__init__.py",
    "*/setup.py",
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

# MyPy configuration
[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
show_error_codes = true

[[tool.mypy.overrides]]
module = [
    "tensorflow.*",
    "talib.*",
    "websockets.*",
    "aiohttp.*",
    "structlog.*",
    "tenacity.*",
]
ignore_missing_imports = true

# Flake8 configuration
[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]
exclude = [
    ".git",
    "__pycache__",
    "build",
    "dist",
    ".eggs",
    "*.egg-info",
    ".venv",
    ".env",
]
per-file-ignores = [
    "__init__.py:F401",
]
